Updated openstack/openstack Project: openstack/nova  49696792a4e2150648a9df523165e9d697de95d0
Clarify max_local_block_devices config option usage

The previous wording for the max_local_block_devices config option made
it sound like if you set it to 0 that you could restrict to only booting
instances from volumes, which is not the case - but is apparently
something that some operators want (this came up in the related bug and
in IRC).

This updates the help text with a description that ndipanov put in the
bug report which makes the 0 case more clear.

Co-Authored-By: Nikola Dipanov <ndipanov@redhat.com>

Related-Bug: #1436314

Change-Id: I01a9e2f193b50285966690efd771da432a7ad40c