Updated openstack/openstack Project: openstack/nova  de989f3c7940372bfbdda890780f064a3fed2e5b
libvirt: use guest as parameter for get serial ports

This is a preparation patch for enabling live migration with
serial console, which will introduce a new user of
get_serial_ports_for_guest().

Move the get_guest call to the caller of
get_serial_ports_for_guest() to avoid the generation of an
exception in a generator.

Change-Id: Ice0a81e576313732d742926c4c05d8251e6ca6d5