Updated openstack/openstack Project: openstack/ironic-specs  0a9e3bd14d2dbb2be42afd6fdf813edabd73391a
Extend the state machine spec to cover DEPLOYWAIT and abort

This patches extend the state machine spec to cover the DEPLOYWAIT state
that was missing and also to talk abort "abort" verb that is being
introduced to the CLEANWAIT state.

The API verb was also corrected, in the API the verb to delete an
instance is "deleted" (for historical reasons) and not "delete".

Change-Id: Id3f7d39027bff43a336cfcdeaf6ba18d6463be51