Updated openstack/openstack Project: openstack/neutron  e11b241a0f46d9a69bdb9396e92aa7df31374daa
ipam: Prevent none from being passed to delete

Previously, it was possible for None to be passed to context.session.delete()
if a port was not found (usually a result of a concurrent delete). This
resulted in an UnmappedInstanceError. This is avoided now by calling
query.delete() directly which does not raise any exceptions.

Change-Id: I3b9f513308d90874ac8e8a7f6345697978d4920f
Closes-bug: #1496554