Updated openstack/openstack Project: openstack/horizon  6822ace491565e3c9d57acd67182ee90ad5b657c
Update JS that adds classes for modal form errors

The JS that processes json error messages for modal forms was not
searching for or adding the correct classes to the HTML for display
the errors. This patch updates the JS so that once processed the
HTML matches the HTML that is returned in the non-json response of
the form.

Change-Id: Ib5ef39d02194ec6c31d69f534831bb0d28e0884c
Closes-Bug: 1443428