Updated openstack/openstack Project: openstack-infra/subunit2sql  d8c555f374054dccc88f11f37a888e056fbd1618
Add --verbose option to show sql statements generated

This commit adds the --verbose option to cli, shell and graph.py.
The deault value is false.  When set, the sqlalchemy.engine logger
is set to INFO which causes the actual SQL Statements to be logged.
Other types of verbose messages could be added in the future.

Change-Id: I1fc0b80dcdd963ae2e72cd313cf6882086da338a