Updated openstack/openstack Project: openstack/python-glanceclient  9f8163ec8207393fc01bdc1717772270197b753a
Don't get the image before deleting it

The client currently downloads the image metadata to check if it's deleted
or not. This logic belongs to the server and it's already implemented
there. Instead of getting the image, send the delete request and catch
the 404 error, which is already raised by the server.

Change-Id: I1e6ef42340f8e380ff99b9d6ca7ea416e0eebfbc
Closes-bug: #1496305