Updated openstack/openstack Project: openstack/openstack-manuals  9e478b0a66f1835c255a95be4423f27a1700e122
Adding developer documentation for openstack/networking-sfc

Adding documentation per [1]. A change has been merged to
add a Jenkins job to generate devref documentation from the
networking-sfc repo. As soon as a change is merged into
the networking-sfc repo it should trigger Jenkins to create
documentation at the URL referenced in this change.

[1] http://docs.openstack.org/infra/manual/creators.html#add-link-to-your-developer-documentation

Change-Id: Ife25ef341b77acb3a8c5e6e3cdfac706f36ce8ea
Depends-On: I1fb10c5281eaa240ce92a58d4f8ba015a2f161b1
Depends-On: I20fb684fc9c8932ec1ea083ff9077e668cc2c1cc