Updated openstack/openstack Project: openstack-infra/storyboard-webclient  76fc5be2ecc9ecf84bd2f986d8ec2761ff2a7cb3
Add paging controls to the result set pager

This commit adds arrows to the result set pager to allow the
user to page through the results of a search[0]. It also adds
a minimal version of the pager, and a list-type attribute.

The "minimal" version fits in a narrow column and has small
text, in order for it to be used on the project_group detail
page.

The list-type attribute is intended to be passed to the
paging functions in the event that there is more than one set
of results to be paged through.

[0]: This is based on an old, abandoned change by Michael
     Krotscheck: https://review.openstack.org/179256/

Change-Id: I3a68937988e0b0dda89f08920fe567d1ce83dbf0