Updated openstack/openstack Project: openstack-infra/storyboard-webclient  4455603319d2513848861b95db5d2fee3bbc6aff
Implement offset/limit pagination of lists

Implement offset/limit pagination of lists in the webclient. This
method of pagination has some drawbacks (for example seeing the same
item on two different pages if new data is added between paging, or
missing out on some items in similar situations), however it is easy
to implement and should be adequate for our use-case for now. Having
some form of pagination is also clearly an improvement on the current
situation.

The `page_size` preference is now only used to determine the page size
of the "main" search result lists, for example /story/list and
/project/list. The lists on the detail pages (eg. "/story/:id" and
"/project_group/:id") have their own page size preference, which is
settable on those pages. This compensates for the difference in page
layout between the lists and the detail pages meaining that different
page sizes are convenient depending on the page being viewed.

Change-Id: I9f511a3c2d0334d393c87afb3fdb4c3b1dca87a8