Updated openstack/openstack Project: openstack-infra/storyboard-webclient  03308d28d46b766c637dac6f41ba1f9b3a6e947d
Change the page title when different things are being viewed

It's useful to have some information in the page title about what is
being viewed on that page. For example, the story title should be in
the page title for a story detail page.

See https://storyboard.openstack.org/#!/story/2000178.

Change-Id: If4f63e90a03ca92f3a92c11c27011233797eb4dd