Updated openstack/openstack Project: openstack-infra/nodepool  c4a74d59e851f686180461c717ac5d0b9a465a45
Correct missing DIB image detection

There are several layers of abstraction:

* A diskimage has a name (eg, "trusty")
* A provider image is the uploaded version of a diskimage.  It
  has its own name, and is also associated with a specific
  diskimage.  Eg, "trusty-check" is supplied by diskimage "trusty".
* A label is a node set made up of one or more nodes launched from
  a provider image.  Its name can also be separate.  Eg,
  "trusty-check-2-node" built from 2 nodes from the "trusty-check"
  image.

The missing diskimage check function was incorrectly looking for
the provider image name in the diskimage table.  Instead, when
it is looking to see whether a built diskimage exists for a given
provider image, it should check the diskimage name for the diskimage
associated with that provider.

The accompanying log message is also updated to be more clear.

Change-Id: I7c5602243b21f4e954b14777c817ffdfb3684e41