Updated openstack/openstack Project: openstack-infra/project-config  bec0a1f5ea38d494b750591adf368d06ff8ae35f
copy_puppet_logs.sh: rename apache config

Rename .conf file to .txt, so we can use the browser to view the files
directly. Currently it's a *.conf.gz file so we need to download the
file, and extract it to see it.

Renaming to .txt.gz will be consistent with other files and will
simplify debug.

Change-Id: I8c6e029124b98190baf94f7d8e7dc179f92bc80d