Updated openstack/openstack Project: openstack/ceilometer  33e4da9464284ec37585bc43265d7be80ef1959a
Refactor keystone handling in discovery manager

The current code assumes the keytone attribute of the manager is a
keystone client object but sometimes this is not true, this is
the exception raised by keystoneclient.

This is not sane to store two differents things in same attribute.
This change fixes that, by storing only the keystone client object and
just reraise the keystoneclient exception when keystone is unavailable.

So the discovery process can detect that keystone is unavailable by
catching the base exception of the keystoneclient library and
just skip discovery plugin with an appropriate message.

Another advantage is that we will not swallow anymore unexpected keystone
client issue by a simple keystone authentication message.

The check_keystone won't work as expected too. Usage of cfg.CONF into a
decorator is broken. This refactor moves the code into the discover
method.

Change-Id: I6adabd7c09672af20e8fabb5016bd2978564e1b2
Closes-Bug: 1493056
Closes-Bug: 1494592