Updated openstack/openstack Project: openstack/neutron  716bc3ba77e60bb365b6783c630401f073568586
Fixes SNAT port not found for internal port

For every router interface added to a router
with a default gateway there will be an internal
SNAT port generated and will be required by the
L3 Agent to process the SNAT rules.

This bug was introduced by the change ID below
Icc099c1a97e3e68eeaf4690bc83167ba30d8099a.

When the gateway is removed these ports have to
be removed from the namespace. These ports are
cached in the router_info and should be provided
to the get_snat_port_for_internal_port function
when called from external_gateway_removed or when
called from _dvr_internal_network_removed.

This patch fixes this problem.

Closes-Bug: #1496578

Change-Id: Id5af4774ba246e24f343f5623af5ea9143bd5f6b