Updated openstack/openstack Project: openstack/python-openstackclient  0116f3abc6c523d721960d838d9dd67535ebc1c2
Support listing users by group name

Listing users within a group is not presently domain scoped. We
do not use the domain info at all when the group option is present.
A new --group-domain option is not needed since we cannot list
users by --project and --group, they are mutually exclusive (as
per the identity API).

Closes-Bug: 1492916
Change-Id: I50f995ee4a03c2bdb21f2b5722546ab8fe786eb6