Updated openstack/openstack Project: openstack/cinder  722b601df94ad7a0721a9bb715c0b0d81507cd88
Correctly report multiattach in Pure drivers

As-is we end up ‘reporting’ to the scheduler that Pure backends do not
support multiattach. This is confusing because volumes created on a
Pure FlashArray will always be capable of doing multiattach even if not
specified by Cinder.

This change syncs up the reporting to the scheduler and what is actually
happening on the array.

Change-Id: I3a1e519efdd0c27f63b049aa686ff9045c260ab7
Closes-Bug: #1493478