Updated openstack/openstack Project: openstack/cinder  3044c70b7eadd5e6de577d9858becf8e2e521179
Remove the unnecassary volume_api.get(context, volume_id)

In the method delete from volume_image_metadata, volume_api.get()
may be called twice. This first time is to get the metadata, and
the second is to get the volume. We can get both of them at the
same time.

Change-Id: I0633c16db21e254ec43bad472748cbc0b7796a24
Closes-Bug: #1489342