Updated openstack/openstack Project: openstack/neutron-vpnaas  d399aa8972e13f79e8c2c5b7de08dccfc6b3dab7
[OpenSwan] Enable usage of the MTU value of an IPSec connection

It is possible to specify the MTU parameter when creating IPSec Site
Connection but this parameter is ignored, because it is missing in
ipsec.conf.template.
This change adds the mtu option to OpenSwan ipsec.conf template.

Refactored existing test_scenario and added a functional test for
OpenSwan driver.

DocImpact

Change-Id: If822454a7acaa3fd003cae3e5e342c8b66ef656c
Closes-Bug: #1478949