Updated openstack/openstack Project: openstack/api-site  2687f751975a124caac9bbb98c3f41bc9c00f96f
Fix a response sample in "Create or update metadata item" API

Even in a sample, a request body and a response body should be
consistent.

In "Create or update metadata item" API, because a value of a
request body is "Bar Value", fix a value of a response body
to "Bar Value".

For details, please see http://developer.openstack.org/api-ref-compute-v2.1.html#setServerMetadataItem

Change-Id: I82a8c50d7454c12c81d95141ceaf1cee5136d128
Closes-Bug: #1496769