Updated openstack/openstack Project: openstack/heat  e3e5a1f3f253b57239f0691fd998964af092853e
Convergence: Work on IN_PROGRESS resources once lock acquired

The check in resource update logic to raise ResourceFailure if a
resource is already IN_PROGRESS does not apply for convergence.

- In convergence, before acting on a resource we try to acquire lock,
hence if a worker is already working on a resource(IN_PROGRESS),
another worker cannot start working on this IN_PROGRESS resource.
- The scenario a worker would reach the above check is when the lock
was forcibly acquired(lets say its worker died). But in this case
under convergence, even if the resource status is IN_PROGRESS we
need to act on it.

Hence removing the above check for convergence.

Change-Id: Ic17a3aa53a94b9e9ab5c841388d509b1f97103c9