Updated openstack/openstack Project: openstack/heat  e84f7e4661d7177e1d8cac6720ee1cb1895f8185
Always return default via template-validate "Default"

Currently we redefine the value of Default when a user value is
provided, which means the data returned doesn't match the schema
defined in the template.  Instead always return the actual default
and add a "Value" key that contains the user value, if provided.

Change-Id: If4ed0b733d4e61c84167063c36cdcb234b001136
Closes-Bug: #1497310