Updated openstack/openstack Project: openstack-infra/project-config  5aa845b1645be7dbef23b22039d55baa4a044d46
Remove usage of transifex client

Remove all invocations of transifex client "tx".

This removes also the script populate-zanata-project.sh that is not
needed anymore, since we have switch to Zanata and started translating
there. The script is broken with the removal of "tx" support in other
places and not needed anymore.

Change-Id: I74350c92da425a6bf32ea1debd217a531dc2f400