Updated openstack/openstack Project: openstack/sahara-image-elements  ba0815ac2f57528903bf2c65126c3e34ea9a8884
diskimage-create.sh: try harder with provides in rpm check

When checking whether a package is installed using rpm, try harder using
the wanted name as provides of a real package.  This fixes the check for
installed python-argparse on CentOS 7, as it is provided by the main
python package.

Thanks to: Brandon James.
Closes-Bug: #1494790

Change-Id: I88d1d0293e3dde9c6087e6e0209f234e922f52ce