Updated openstack/openstack Project: openstack-infra/storyboard-webclient  ec0f5651cd1225915e0b09cefb13f3979dd5c86e
Add 'Remove this story' button to detail page

This adds a 'Remove this story' button next to the 'edit' and
'subscribe' buttons on the main story detail page. Previously, it was
only possible to remove stories from the edit form, which was not the
first place a new user would look.

Change-Id: I55aa9c5b6184169890bc1e4f38ba08b3e35d676b