Updated openstack/openstack Project: openstack/ironic  652e9a97098bcf833df0e31a3c6f79fb2d59f553
Handle unquoted node names from virt types

Not all virt types return node names in quotations when requesting a list of running VMs.
This should not be a requirement by the code, so relax this restriction and match
node names without quotes.

Also, add comments to clarify the expected return value for each of the commands.

Change-Id: I55a259f8687627a4ebd2d8bb941eb5b1041045c7
Closes-Bug: #1496867