Updated openstack/openstack Project: openstack/barbican  e2de47e18469d5f807e597b0cd60d303eaa20195
Add missing X-xxxx HTTP headers to the unauth context

The unauth context is used by Barbican out of the box for local API evaluation
as well as for deployments where an authenticating proxy (such as Repose) is
placed in front of Barbican. Hence this unauth context must provide all the
required HTTP header settings to the Barbican context prior to policy engine
computation for RBAC and so forth. This CR provides those missing headers.

Change-Id: Ib1d6abdb384f5883aa06ec1b6811c3ad80291568
Closes-Bug: #1497639