Updated openstack/openstack Project: openstack/heat-templates  112fc75a9174c2e623fd28953cea392ab9d4eab0
Add puppet hook option for --tags

Using --tags allows you to run a puppet manifest and only apply
a subset of the catalog. You can specify --tags package for
example to only install packages resources for a given
manifest.

The intent is to use this option alongside of the TripleO docker
work to selectively control what parts of a puppet manifest
get executed in a container.

Change-Id: Ic8b54c89e62cbe5074f256e29e10f59aa2ab0aae