Updated openstack/openstack Project: openstack-infra/nodepool  8f3db8930666ec3b82de53b18a97c438b8487ef7
Decouple nodepool db and config from builders

Currently we use the nodepool config object and nodedb database
to determine information about the diskimages we build. Start parsing
the config on our own and stop using nodedb.

This changes images to use a uuid for ID's since the primary key
ID which was previously used is no longer readable by the builders.

Co-Authored-By: Joshua Hesketh <josh@nitrotech.org>

Change-Id: Id1bb431839cde2bd7b42cf1f57460febd56849b3