Updated openstack/openstack Project: openstack/manila  a8e5154cf30119d64981b7aba2246abbb681492b
Fix error response when denying snapshot creation

If we try to create snapshot from share that is not snapshottable
we expect error that it is not possible and proper description for it.
Now we get some abstract message
"ERROR: Unable to process the contained instructions".

So, make API provide proper information about error.

Change-Id: I16a2f7c80102159c30f91af7e045e73ee3df7f89
Closes-Bug: #1496408