Updated openstack/openstack Project: openstack/nova  c96ade8eb38387af962f6cf307f20f317a7729fd
claims: move a debug msg to a warn on missing migration

We want to log a warning when there is a missing migration record when
creating a migration_context record, because it's most likely caused by
either a misconfiguration or a bug. We made it a debug message in the
original patch as the rules for translated strings were unclear at the
time, but now we can do it - so let's do it!

Change-Id: I440e357f8eeacae2972c74aa7be15d689fb9ac64