Updated openstack/openstack Project: openstack/neutron  2845d5fb5060dab1246d78e49bdfe7cea23cfd07
Fix adding tap failure if bridge mapping is not provided

When bridge mapping option is not provided, tap interface cannot be
added properly. The root cause is that linux bridge agent is expecting
bridge mapping option is provided if physical network is presented.

However it is not always the case as Kilo was working fine
without existing bridge mapping

Change-Id: Ie43905887a9bc60858e3edae7fc339a7e0485716
Closes-bug: #1495161