Updated openstack/openstack Project: openstack/ironic  38a4ebcdf8cf1c8b47bf1f9bd32e4d55b5e31a43
Allow abort for CLEANWAIT states

This patch allows a node in CLEANWAIT to be aborted.

The @clean_step decorator has been extended to accept a new parameter
"abortable". By default this parameter defaults to False since a clean
step could potentially brick a machine so we better make it explicit.

If the clean step is abortable, the process of aborting will happen
immediately; if the clean step is not abortable the abortion will happen
as soon as the clean step is done. If a clean step is marked to have
the abortion done after its completion but it is the final clean step in
the cleaning operation the cleaning process will just finish successfully.

A new verb 'abort' is being added to the API and the microversion is
being bumped to 1.13.

Closes-Bug: #1455825
Change-Id: Ia6846c048b3dab44a8280366a7305aca1d3eb783