Updated openstack/openstack Project: openstack/heat-templates  a288b370f516bf6b0a1ce711770f6f6c66793883
Fixes zuul failure

template-validate is now updated to validate the template
by using stack create validation path, which causes some of
the existing templates in this repository to fail.
so this patch is added to mark them as invalid and skip them
as part of template validation.

It was failing becuase of circular reference among the resources
in these invalid templates.

depends-on: I15891196be94d1e100869edeba0a7ef557edb36a
Change-Id: I6a003f35495bb43e0ca2a824eb9e503c4059c37a