Updated openstack/openstack Project: openstack-infra/subunit2sql  5af2ca3f8bdbb60abcbfc823872fb9ef42891d7c
Add migration to fix wrong column name in test_metadata

This commit adds a new migration to fix an incorrectly named column
in the test_metadata table. The test_id column was accidentally named
test_run_id, likely a copy and paste issue from the test_run_metadata
table. Since this table was mostly unused up until this point the
typo was never noticed. However, all the db api classes and methods
refer to the column as test_id, so when you try to interact with
test_metadata it fails.

Change-Id: Ib722a318f970e0f5d194b81cbb0f196d772b956d