Updated openstack/openstack Project: openstack/neutron  ec9a634893eb1acafa42b4a4676f9e37789bbdba
ml2: don't consider drivers with no bind_port for qos supported rule types

If the driver is not going to bind a port, we should not use it as
justification to limit the subset of rule types available for the user.

The most common scenario for that would be l2pop, that does not handle port
binding, but merely modifies the implementation details of node
interconnection.

Change-Id: Iba0f3530deff3bb2871ebf893c43ef628a80c7c0
Closes-Bug: #1488996