Updated openstack/openstack Project: openstack/oslo.db  7daffffafc8940a5097cdd33724eff1efaa33d1a
Add support for pickling enginefacade context objects

In order to pickle a context object which refers
to an active _TransactionContextTLocal, we have to
use __reduce__ so that no database, engine, or session
state is pickled, and only a constructor remains.  Upon
unpickling, the constructor will install a new, empty
_TransactionContextTLocal on the new context, as is
appropriate since this is a totally different context object
than the original one.

Change-Id: Ia50cf10ff91a013ee24773c095a3df69ae06847b