Updated openstack/openstack Project: openstack/sahara-image-elements  4d66eedd4c756c491b939577fcccb1db07fafe3c
hadoop: move scripts from post-install.d to install.d

As also the previous comment on the top of 40-setup-hadoop said, this
was done because the Hadoop v1 RPM was buggy, so it needed to be
installed after the rest of the system is set up.

Since the support for Hadoop v1 has been dropped, the installation of
Hadoop can be done in install.d.

Change-Id: If3b0f8f595d6bf36017e63d331b5c1f7faa532e2